会員登録継続、他人削除しないの途 中

+++ recipe.php
@@ -10,20 +10,24 @@
     <?php
     require('library.php');
     $db = dbconnect();
+    session_start();
 
-    $stmt = $db->prepare('select * from recipen where id =?');
+    $stmt = $db->prepare('select r.id, r.recipename, r.member_id, r.image, r.foodstuffs, r.recipe, r.created,  
+    r.modified, m.name from recipen r, member m where r.id=? and m.id=r.member_id');
     if (!$stmt){
-        die($db->error);
+    die($db->error);
     }
+   
     $id = filter_input(INPUT_GET, 'id', FILTER_SANITIZE_NUMBER_INT );
     $stmt->bind_param('i', $id);
     $stmt->execute();
 
-    $stmt->bind_result($id, $recipename, $member_id, $image, $foodstuffs, $recipe, $created, $modifind);
+    $stmt->bind_result($id, $recipename, $member_id, $image, $foodstuffs, $recipe, $created, $modified, $name);
     $stmt->fetch();
 
     ?>
-    
+    <div><?php echo h($name) . 'さんのレシピん♪' ; ?></div>
+
     <div><?php echo h($recipename); ?></div>
     <time><?php echo h($created); ?></time><br>
     <img src="recipe_picture/<?php echo h($image); ?>">
@@ -33,11 +37,18 @@
     <div><pre><?php echo h($recipe); ?><pre></div>
 
     <div>
-        <a href="update.php?id=<?php echo $id; ?>">編集する</a>|
-        <a href="delete.php?id=<?php echo $id; ?>">削除する</a>|
-        <a href="toppage.php">TOPページに戻る</a>
+            <a href="toppage.php">TOPページに戻る</a>
     </div>
-  
-    
+
+    <?php 
+        $clear = '';
+        if (isset($_SESSION['id']) && isset($_SESSION['name'])){
+            $clear = 'clear'; 
+        }
+    ?>
+  <?php if ($clear == 'clear'): ?>
+    <a href="update.php?id=<?php echo $id; ?>">編集する</a>|
+    <a href="delete.php?id=<?php echo $id; ?>">削除する</a>|
+  <?php endif; ?>
 </body>
 </html>
